requirements.txt
paho-mqtt==1.4.0
flask==1.1.1
flask-bootstrap==3.3.7.1
flask-socketio==4.2.1
flask-sqlalchemy==2.4.0
PyMySQL==0.9.3
sqlalchemy==1.3.6
